[Refactor] seungwon: 기존 컴포넌트 프리뷰를 component-test.tsx로 변경 및 디렉토리 정리 + App.tsx 수정
